Fixed all the FindBugs warnings

+++ technologies/BasicESB/src/java/com/pjaol/ESB/monitor/MonitorBean.java
@@ -45,7 +45,7 @@ public class MonitorBean {
 		if (c == 0)
 			return 0;
 		
-		return (t / c);
+		return ((double) t / c);
 	}
